Check if the process instance is actually waiting for the 'ResponderAvailable' signal before signaling.
Under load, it had been observed that the current rhpam transaction
sometimes is not yet committed by the time the 'ResponderUpdatedEvent'
event meage is consumed from the Kafka queue. When this happens the
signaling fails and the process instance is stuc in the signaling node.
To workaround this issue, a query is performed against the rhpam
database to ensure that the process is actually waiting for the
'ResponderAvailable' signal before signaling the process. If the process
instance is not waiting on the signal - which means the rhpam
transaction is not yet committed - the thread pauses for 300ms, and the
query is executed again, up to a maximum of five times.